Protect against creating double packing lists.
Whenever software instance is begin started or destroyed new packing list have
to be created. But in both cases it is not desirable to have more then one
active packing list related to current step of state. So dissallow it by
passing tag and checking for its existence. Even more -- serialize in order to
lock operation on one Software Instance in transaction.