Reverse the order in which repositories are added to sys.path.
Detailed rationale:
erp5/softwre.cfg defines
  repository_id_list = erp5
It is not possible to prepend a value to another in buildout configuration
file syntax, but provides a way to append:
  repository_id_list += foo
In such case, when running tests, one will want to access the test(s)
folder project's repository. But that folder will be shadowed by erp5's.
There is no trivial fix without drawbacks, and this fix is trivial enough
to be considered better than other fixes.