Fix bridge issue with Linux >= 2.6.39 and bridge TAP-only interfaces.
In order to  be able to check  the uniqueness of IPv6 address  assigned to the
bridge,  the  network  interface  must  be up  from  an  administrative  *and*
operational point of view.

However, from  Linux 2.6.39, the bridge  reflects the state  of the underlying
device (e.g.   the bridge  asserts carrier if  at least  one of its  ports has
carrier) whereas it always asserted  carrier before. This should work fine for
"real" network interface, but will not  work properly if the bridge only binds
TAP  interfaces, which,  from  2.6.36, reports  carrier  only and  only if  an
userspace program is attached.