Added new parameters to Software init method.
Added upload_cache_url and upload_dir_url:
 - upload_cache_url: the value must be the url of shacache used to
   upload
   content.
 - upload_dir_url: the value must be the url of shadir used to upload
   the
   information about the content.

Also, the signature_private_key_file is not required to instanciate a
Software object, then it can be None by default.

All these parameters listed above should only be used in the buildout
command line, if they are not None. And it should be defined under
networkcache-section.