Now it is possible to define the section networkcache into slapos.cfg.
Inside of slapos.cfg you can define the section called networkcache,
which should contain all the required configuration to upload files
to shacache and shadir.

Options:
  - upload-cache-url: url to upload content to shacache
  - upload-dir-url: url to upload content to shacache
  - signature-private-key-file: private key file used to sign the
    information which may be sent to shadir

Note: the option signature-private-key-file, may also be used as
parameter of slapgrid command.
   (i.e. --signature_private_key_file=/path/to/private.pem)

And of course, following the previous commit, we need to propagate these
parameters to Software class.