Force lxml-python installation.
lxml-python is part responsible for creating develop egg, which is build with
specific configuration.

As this profile is using another hack: develop eggs which does not exists,
force (hopefully) to build lxml python as first part in order to not have it
fetched and build in uncontrolled way by slapos.cookbook begin locally
developed.