Skeletonise PAS configuration.
It follows idea of alarm based promises which can detect system misconfiguration.

The reason to go this way, as the explanation of what to do and what to check
is becoming really complex and would take more time and space to write then
simple code which will stabilise site.

The test shall use this alarm to configure PAS, so it will be known that it works.

Until scripts are implemented alarm is disabled.