Move sub-partitions parameter down a nesting level.
Simplifies editing the software release by removing the need to edit
instance-erp5-cluster.cf.in each time.
Applied to kumofs, cloudooo, mariadb and zodb.
Partially applied to frontend, as it is more complex.
Not applied to zope, as it is even more complex - but also topical to this
software release.
Also, change default timezone to UTC, to be neutral.
Also, reformat section extension.