Check component's binary creation.
Add a new section ["component"-output] to test the creation of the expected
build output.
The verification is simply done with the "test" command.

The output file are reference as section variable, which allow to reuse the
location in the software profile, and so, prevent harcoding the local directory
path in the software profile.