Slapos_accounting: Always create income, receivable and collected_vat in init script
erp5_accounting only create income , receivable and collected_vat in SaleInvoiceTransaction_init
if Invoice Line is not in visible allowed content type. But the method getVisibleAllowedContentTypeList seems
to not work well and always return 'Invoice Line' in the list event after adding 'Invoice Line' in
hidden content types.
To fix this for SlapOS, we create a custom SaleInvoiceTransaction_init in slapos_accounting which always create
income, receivable and collected_vat.