logrotate: delay compression of rotated files
Most of the time, the logging process keeps the log open, accesses it via
a file descriptor and gets notified of log rotation via a signal. Because
logrotate does not wait that the log is close (could it do it with inotify ?),
there can be race conditions between the compression and events that are logged
at the same time.

In the case of NEO, logs are SQLite DB. Rotation is also notified via a signal
and VACUUM is sometimes called to reduce its size.