erp5: review request parameters for SLA & ZODB
All parameters about SLA, i.e. computer-guid & instance-guid, are removed in
favor of a new "sla-dict" parameter, which is easier to implement and much more
versatile.

All changes in the request parameters are incompatible. The old ones are
ignored without warning/error.

For compatibility, the reference of ZEO partition is still "zodb".

Default settings were also fine. Default name of mount-point and FileStorage
file is reverted to 'root' instead of 'main'.