slaprunner: fix slapos-software parameter usage
It was applied by the combination of [slapos-repo] that checkouts master
branch and [slapos-repo-config] that checkouts
$${slap-parameter:slapos-software} branch afterwards.
When instance profile runs for the second time, [slapos-repo] resets to
master and [slapos-repo-config] does nothing.
This fix consists in using directly $${slap-parameter:slapos-software}
in [slapos-repo].

[slapos-repo-config] was also setting up .project, this is still the
case and the section was renamed to [instance-software] for
consistency.