Allow to create tap and nat interface
A new parameter 'use-nat', True by default will specify if we will setup a nat interface.
if 'use-tap' is Truei (false by default), a tap interface will be connected with a new interface in kvm.
if 'use-tap' and 'use-nat' are True, then virtual machine will have two interfaces, one with tap and
one with nat.