erp5: Use the same default value for thread amount as Zope's.
One thread is just too limited to produce a usable instance (for example,
a single-zope setup with default thread count fails to run live tests). It
is also not an advised setting for activity nodes, because of timerserver
still queuing process_timer calls while CMFActivity is busy processing; the
result being a (limited) waste of memory, and hammering mysql after long
processing periods.
Zope's default value must be duplicated because thread-amount is also used
to generate haproxy configuration file (so just not providing the value in
Zope's configuration file is wrong).