kvm: can now accept parameters external-disk-format, cpu-options and numa
kvm_run script can now create external disk with specified format (qcow2, raw, etc...)
if kvm external disk is created it will never be replaced, even if user
reduce the number of disk with parameter 'external-disk-number' or change the
'external-disk-format' value. All created disk will be added to qemu run command.

numa option can be used to create [fake] numa nodes and expose them to the guest
OS instance. In this case it is important to know the host cpu topology and where
the instance will be allocated.

cpu-options are additional parameters like sockets=1,cores=5,threads=1
in this case if cpu-count parameter is 1 qemu options will be:
-smp 1,sockets=1,cores=5,threads=1