openssl: Convert patches to -p1
We'll soon need to apply/ship external patches for openssl with us, and
external patches and patches in general usually go in -p1 format.

Since patch-options is "global" - i.e. it affects all patches for
slapos.recipe.cmmi, this change is needed for us to keep
downloaded-from-somewhere patches as unmodified as possible.

/cc @kazuhiko