wendelin.core: It needs git at runtime for wendelin.core-dev
Because git is used in `python setup.py sdist`

    https://lab.nexedi.cn/nexedi/wendelin.core/blob/11d130d1/setup.py#L168

we need to make sure it is available at runtime (at time when sdist
happens), or else installation could fail.

Since zc.recipe.egg:develop now supports setting environment variables

    https://lab.nexedi.cn/nexedi/slapos.buildout/commit/93d35f5d
    https://lab.nexedi.cn/nexedi/slapos.buildout/commit/71078537 )

and we use the proper version of it in slapos stack:

    https://lab.nexedi.cn/nexedi/slapos/commit/d0807d90

we can just set environment with git being in PATH and fix it this way.

/cc @Tyagov, @kazuhiko, @jerome
/reviewed testbot(building ok), trustme