Wendelin: Add jupyter(ipython-notebook) to wendelin software release
Now we have a combined software release for wendelin(erp5-cluster) and ipython-notebook(jupyter).
The main instance wendelin will request for notebook as well as erp5-cluster.
Different software-types available are :-
1. default - Will behave exactly same as default for erp5-cluster. Return parameters for various softwares in erp5-cluster.
2. create-erp5-site: Will create erp5-bootstrap service and start erp5.

In the parameters, users have to provide with "jupyter-family" key whose value would be the family user wants jupyter to use while interacting with erp5.