switch-softwaretype: make it possible to call it in chain
Store only first 2 system arguments in 'args' so that to facilitate use of switch-softwaretype recipe in chain.
For example:-
In one buildout having section with recipe switch-softwaretype and running buildout for a file which in succession is calling another buildout, so that we don't have a conflict between other system arguments.