Revert "usersyslog: New component."
This reverts commit d48851e977e2c62b5ef35a3fa914f36d21aca063.

Turns out, "logger" command (with which this tool was developped) is not
actually using libc's syslog API, so is not representative of proper
syslog usage. Also, syslog API happens to be extremely painful to hook
(500 lines of libc code to fork because of direct internal accesses,
constants and static globals).
Syslog hooking will happen someday, just not this way.