Jupyter: Add display_response parameter and rename response param to message
'display_message' parameter would be used to check if there is a need to display 'message'
after making a successful request to erp5 or not.

Useful for magics such as '%my_notebooks' for which we don't need to show message cause we are
already getting some content from erp5 for this magics.