component/postgresql: Factor out common build stuff to postgresql-common
Currently we have recipes to build both postgresql 9.1 and 9.2 (see
9f1f0759 "provide both postgres 9.1 and 9.2") which mostly duplicate
each other with minor difference that 9.1 is built with perl and 9.2
without (perl added to 9.1 in dbbd9a96 "Include Perl in Postgres").

To reduce the duplication let's move common compilation bits to common
section.

NOTE I've tried to add perl to postgresql 9.2 as well and got the
following compilation error:

    ld: .../perl/libs-c/libperl.a(op.o): relocation R_X86_64_32 against `.rodata.str1.8' can not be used when making a shared object; recompile with -fPIC
    .../parts/perl/libs-c/libperl.a: could not read symbols: Bad value

which happens because libperl.a is not compiled with fPIC.

For now we don't need perl in postgresql92 and this is not handled, and
we just deduplicate building recipes without any actual change for
resulting commands how the software is built.