redis: v↑  (2.8.23)
- update Redis software to latest upstream in 2.8.* series (which now
  supports IPv6 out of the box);

- update Redis instance template to the one from 2.8.23 and re-merge our
  templating changes to it (file/dir locations, port and binding, master
  password).

  The whole diff to pristine 2.8.23 redis conf is now this:

  diff --git a/.../redis-2.8.23/redis.conf b/slapos/recipe/redis/template/redis.conf.in
  index 870959f..2895539 100644
  --- a/.../redis-2.8.23/redis.conf
  +++ b/slapos/recipe/redis/template/redis.conf.in
  @@ -46 +46 @@ daemonize no
  -pidfile /var/run/redis.pid
  +pidfile %(pid_file)s
  @@ -50 +50 @@ pidfile /var/run/redis.pid
  -port 6379
  +port %(port)s
  @@ -69,0 +70 @@ tcp-backlog 511
  +bind %(ipv6)s
  @@ -108 +109 @@ loglevel notice
  -logfile ""
  +logfile %(log_file)s
  @@ -174 +175 @@ rdbcompression yes
  -# hit to pay (around 10%) when saving and loading RDB files, so you can disable it
  +# hit to pay (around 10%%) when saving and loading RDB files, so you can disable it
  @@ -192 +193 @@ dbfilename dump.rdb
  -dir ./
  +dir %(server_dir)s
  @@ -217 +218 @@ dir ./
  -# masterauth <master-password>
  +%(master_passwd)s

NOTE There are test failures for almost all Redis versions when machine
have not small amount of CPUs:

    https://github.com/antirez/redis/issues/2715#issuecomment-151608948

Because the failure is in replication test, and so far we do not use
replication, and there is no feedback from upstream author to handle
this (for 7 days for my detailed report, and for ~ 3 months for this
issue in general), we can just disable replication test as a temporary
solution.  ( to handle remote patches with md5 hash easily the building
recipe is changed to slapos.recipe.cmmi )

NOTE Redis updated to 2.8 version because GitLab uses this series.
If/when we need more recent one we can add [redis30] in addition to
[redis28].

/cc @alain.takoudjou