helloworld: Unify naming for internal "hello web" service
Since 7cd2f953 (helloworld: Turn it into real web-service) helloworld SR
implements real web-service which talks to outside world with the help
of shipped-with-it "hello web" program.

Buildout sections which create helloweb service and promise were named
based on helloworld prefix though, which is not very consistent. Thus the
change for such sections to the same name as the "hello web" program uses.

Also for style consistency the "hello web" program is renamed from
hello-web to helloweb, as this naming style is usually used for
helloworld.{c,py.rb,go,etc}

So the final naming is:

    helloweb

for both program and related sections.