helloworld & helloweb: Ruby version
To illustrate how to build Ruby software, let's add helloweb-ruby
component and integrate it into helloworld SR as a service.

helloweb-ruby installation is a bit tricky, because we do not have it
initially as a released gem, and have to build it with Bundler by hand.

Changes for integration into helloworld services infrastructure are
small, thanks for the "kind" rework we did in previous patch.

/cc @kazuhiko