Go component
Add support for Go language in the form of providing component for Go
toolchain and stdlib.

Current Go stable series is 1.5.* but to bootstrap it one need to have
Go 1.4.* available, which itself needs C compiler to bootstrap.

We need Go support for upcoming GitLab SR, where one server is written
in this language:

    https://gitlab.com/gitlab-org/gitlab-workhorse

/cc @gabriel, @Camata