gitlab/gitlab-shell-config.yml: Explicitly point it to secret file
Explicitly point gitlab-shell to location where we keep secrets.

We already pointeg gitlab to that place and now we do that for
gitlab-shell so those 2 peieces can connect to each other ok.

Regarding the setting itself - there is no such block in omnibus-gitlab,
but it is present in gitlab-shell configuration example:

    https://gitlab.com/gitlab-org/gitlab-shell/blob/82b3a4e8/config.yml.example#L35

/cc @kazuhiko, @jerome